+++ tacacs-plus-protocol/src/accounting/owned.rs
@@ -1,11 +1,12 @@
-use std::borrow::ToOwned;
 use std::string::String;
+use std::string::ToString;
 
 use super::{Reply, Status};
 use crate::owned::FromBorrowedBody;
 use crate::sealed::Sealed;
 
 /// An owned version of a [`Reply`](super::Reply).
+#[derive(Debug, Clone, PartialEq, Eq, Hash)]
 pub struct ReplyOwned {
     /// The status returned by the server.
     pub status: Status,
@@ -25,8 +26,8 @@ impl FromBorrowedBody for ReplyOwned {
     fn from_borrowed(borrowed: &Self::Borrowed<'_>) -> Self {
         ReplyOwned {
             status: borrowed.status,
-            server_message: borrowed.server_message.as_ref().to_owned(),
-            data: borrowed.data.as_ref().to_owned(),
+            server_message: borrowed.server_message.to_string(),
+            data: borrowed.data.to_string(),
         }
     }
 }