+++ tacacs-plus/src/response.rs
@@ -2,7 +2,7 @@ use tacacs_plus_protocol::Argument;
 use tacacs_plus_protocol::{authentication, authorization};
 
 /// The final status returned by a server during a TACACS+ session.
-#[derive(Clone, Copy, PartialEq, Eq, Debug)]
+#[derive(Clone, Copy, PartialEq, Eq, Debug, Hash)]
 pub enum ResponseStatus {
     /// The operation succeeded.
     Success,
@@ -12,7 +12,7 @@ pub enum ResponseStatus {
 
 /// A server response from an authentication session.
 #[must_use = "Authentication failure is not reported as an error, so the status field must be checked."]
-#[derive(PartialEq, Eq, Debug, Clone)]
+#[derive(PartialEq, Eq, Debug, Clone, Hash)]
 pub struct AuthenticationResponse {
     /// Whether the authentication attempt passed or failed.
     pub status: ResponseStatus,
@@ -26,7 +26,7 @@ pub struct AuthenticationResponse {
 
 /// A TACACS+ server response from an authorization session.
 #[must_use = "The status of the response should be checked, since a failure is not reported as an error."]
-#[derive(PartialEq, Eq, Debug, Clone)]
+#[derive(PartialEq, Eq, Debug, Clone, Hash)]
 pub struct AuthorizationResponse {
     /// Whether the authorization attempt succeeded.
     pub status: ResponseStatus,
@@ -42,7 +42,7 @@ pub struct AuthorizationResponse {
 }
 
 /// The response from a successful TACACS+ accounting operation.
-#[derive(Debug, PartialEq, Eq, Clone)]
+#[derive(Debug, PartialEq, Eq, Clone, Hash)]
 pub struct AccountingResponse {
     /// The message that can be displayed to the user, if any.
     pub user_message: String,