+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 #### Added
 
 - CI tests are also run against [TACACS+ NG], an actively maintained TACACS+ server implementation (#30)
+- Common std trait implementations (e.g. `Hash`, `PartialOrd`/`PartialEq`, `Debug`, `Display`) to publicly exposed types
 
 #### Changed
 
@@ -27,10 +28,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `FieldText::from_string_lossy()` constructor that automatically escapes any non-printable-ASCII characters (#31)
 - `InvalidText` type, for `FieldText` construction errors
+- Common core trait implementations (e.g. `Hash`, `PartialOrd`/`PartialEq`, `Debug`, `Display`) to publicly exposed types
 
 #### Changed
 
 - `FieldText`'s `TryFrom` & `FromStr` implementation error types were changed to `InvalidText` (#35)
+- `authentication::Action::SendAuth` is no longer marked as `#[deprecated]`, since [RFC8907 section 10.5.3] only recommends against
+  its use, not deprecates it
+
+[RFC8907 section 10.5.3]: https://www.rfc-editor.org/rfc/rfc8907.html#section-10.5.3-4
 
 ## [0.2.2] - 2024-08-20